fix: 🐛 map randomized options to block.options correctly
Before the indices of selected options were taken from the randomized
block.options array. Hence, the selectedOptionIndices were only correct
with the original order of block.options. This resulted in different
selected options when the presenter was initialized with an initalState
containing selectOptionIndices (since a new random block.options order
was created).